+++ src/constants.ts
@@ -13,6 +13,7 @@ export enum Asset {
   JTO = "JTO",
   ONEMBONK = "ONEMBONK",
   SEI = "SEI",
+  JUP = "JUP",
   UNDEFINED = "UNDEFINED",
 }
 
@@ -46,6 +47,7 @@ export const FLEX_MINTS = {
       "LF8rzt8WSzMuH3E4pfryGvSWMc1jDponXdQHkzL2cGo"
     ),
     [Asset.SEI]: new PublicKey("Dju5Nvuet6GyBk7dBwWSG9cg2JDhtxKXHsBTs7HFbPk3"),
+    [Asset.JUP]: new PublicKey("rYFexooTE27wUDaJsjpy8UdyKqUHxroPvfB2Hgmajjt"),
   },
   devnet: {
     [Asset.APT]: new PublicKey("FbfkphUHaAd7c27RqhzKBRAPX8T5AzFBH259sbGmNuvG"),
@@ -63,6 +65,7 @@ export const FLEX_MINTS = {
       "76x829V8cNWymEBNjUuE22bUcVnShNeRwnXnegviejyj"
     ),
     [Asset.SEI]: new PublicKey("CTw2xSSAfrv9hJGVpB2R2q5xYrdX79i3hXeCiQiAKf2f"),
+    [Asset.JUP]: new PublicKey("7uP5h7kxRaSbd2dz5e6gZp8yhqazyMvaXVoNZ4HsgZ2n"),
   },
 };
 
@@ -167,6 +170,7 @@ export const PYTH_PRICE_FEEDS = {
       "G6PKiDSNpyeFRM6iLyZWXTP4U8sKejdEHpDGEXJxTpdP"
     ),
     [Asset.SEI]: new PublicKey("EhC74qLWnp3i5y3EiPibQHdAkNU4F9kvcDuWgcQuLj6G"),
+    [Asset.JUP]: new PublicKey("8uMQw668EowuxXcX6oYohLepTCoZmk9WEnj224qevLqj"),
   },
   devnet: {
     [Asset.SOL]: new PublicKey("J83w4HKfqxwcq3BEMMkPFSppX3gqekLyLJBexebFVkix"),
@@ -192,6 +196,7 @@ export const PYTH_PRICE_FEEDS = {
       "8ihFLu5FimgTQ1Unh4dVyEHUGodJ5gJQCrQf4KUVB9bN"
     ),
     [Asset.SEI]: new PublicKey("6cUuAyAX3eXoiWkjFF77RQBEUF15AAMQ7d1hm4EPd3tv"),
+    [Asset.JUP]: new PublicKey("g6eRCbboSwK4tSWngn773RCMexr1APQr4uA9bGZBYfo"),
   },
 };
 
@@ -216,8 +221,8 @@ export const ACTIVE_MARKETS = ACTIVE_EXPIRIES * PRODUCTS_PER_EXPIRY + 1; // +1 f
 export const TOTAL_EXPIRIES = 5;
 export const TOTAL_MARKETS = PRODUCTS_PER_EXPIRY * (TOTAL_EXPIRIES + 1);
 export const PERP_INDEX = TOTAL_MARKETS - 1;
-export const ACTIVE_PERP_MARKETS = 11;
-export const UNUSED_PERP_MARKETS = 14;
+export const ACTIVE_PERP_MARKETS = 12;
+export const UNUSED_PERP_MARKETS = 13;
 
 export const DEFAULT_EXCHANGE_POLL_INTERVAL = 30;
 export const DEFAULT_MARKET_POLL_INTERVAL = 5;
@@ -373,11 +378,11 @@ export const STATIC_AND_PERPS_LUT: {
     },
   }),
   mainnet: new AddressLookupTableAccount({
-    key: new PublicKey("DSzLR6CsMTKB3AmUVKSWw6dHdGtuqmryD9gsXewNwt7A"),
+    key: new PublicKey("8EUkL6ybmsn3br8TTUiJBYoSZuivD6V9Sb7Gcas9m164"),
     state: {
       deactivationSlot: BigInt("18446744073709551615"),
-      lastExtendedSlot: 241132987,
-      lastExtendedSlotStartIndex: 140,
+      lastExtendedSlot: 245012318,
+      lastExtendedSlotStartIndex: 160,
       authority: new PublicKey("BR47Yncz7AKKD4f6NuP83yxb31YjvYvZuHApjEvtP8tu"),
       addresses: [
         new PublicKey("8eExPiLp47xbSDYkbuem4qnLUpbLTfZBeFuEJoh6EUr2"),
@@ -408,6 +413,8 @@ export const STATIC_AND_PERPS_LUT: {
         new PublicKey("8ihFLu5FimgTQ1Unh4dVyEHUGodJ5gJQCrQf4KUVB9bN"),
         new PublicKey("11111111111111111111111111111111"),
         new PublicKey("6cUuAyAX3eXoiWkjFF77RQBEUF15AAMQ7d1hm4EPd3tv"),
+        new PublicKey("11111111111111111111111111111111"),
+        new PublicKey("g6eRCbboSwK4tSWngn773RCMexr1APQr4uA9bGZBYfo"),
         new PublicKey("11111111111111111111111111111111"),
         new PublicKey("JE6d41JRokZAMUEAznV8JP4h7i6Ain6CyJrQuweRipFU"),
         new PublicKey("EaNR74nCjrYyNDsuoWmq19pH76QSd1nuTzvJSr3RDQ6x"),
@@ -530,6 +537,17 @@ export const STATIC_AND_PERPS_LUT: {
         new PublicKey("VdYom29m1yXemVTYFm1cw8Ycu7yzScxyzvG4P63wfW2"),
         new PublicKey("9Lsei9qMonizuuusFgdh94djt7fpgktGmh9PRi2j2GUQ"),
         new PublicKey("EU9uvnDuqNgunfmnxNAjhYi8iEKV3m7gnr5vatg2tgPj"),
+        new PublicKey("3F4cdKDeLpvnUtxJ45ue4gs5kCZmhaeLDXN77RHGPfHu"),
+        new PublicKey("GkjperyMoy6MCUyg73KzZy4WbfaRhKYAoTBjW3JRjYuG"),
+        new PublicKey("8KTa6xFc1dTpm7fsSDKMe5zftHzC6mCR38f7XUp6ugqs"),
+        new PublicKey("2iuKYW8qnPPjeWn9rpJUdJQvNWM2CyiSFZNprUFtcYQm"),
+        new PublicKey("2FEs2ttf6vauffHDhvWf6F3HANB26iiSULxNF3FAdW1d"),
+        new PublicKey("5raajUcFKgaRbyh6yRkAKWroLe6PwtgCAU1vT7wmmpFu"),
+        new PublicKey("7Hz86pfpm5hxmzuJbNFbfJ8LEU8EnLL7fc3oyFgVrm5G"),
+        new PublicKey("FYpmTRddSQufU9ZrwPHgAj39epj4dP3Qwfdznm3oC5SY"),
+        new PublicKey("jQdzjLTZg7fQWGuWHjZmrkTjDNQ3DrAVKkzWvXVrzRm"),
+        new PublicKey("DdFzaannSyXs12sB517GbtCsZihh8X9RS2ZKJKuTp5sG"),
+        new PublicKey("2Vaf7yEBXi86SGhDM79w1dRQxz5NC2YFWshm5NxkRUDB"),
       ],
     },
   }),